Use 2-pin horizontal SMD PicoBlade MX 1.25mm battery connector.
The inaccurate term JST 1.25mm, or 'Micro JST 1.25mm' is often
used for these. The connector socket is under 4mm so should fit,
and this is commonly offered as the connector with LiPo batteries.

Putting this on both sides in place of the original (harder to
find) connector and soldering points.