Enlarge most 2x2.6mm hotswap pads to 2.6x2.6mm
This is more symetric and more closely matches the original Forager
hotswap footprint. Unclear why the Ceoloide Ergogen hotswaps are
that small on one side - although they work fined fine for me as is
on another board.

A few are still on the small side due to constaints like the battery
cut out, or wiring.